Fixed some warnings in libwin

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@25337 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/windows/lib/registry.c
@@ -116,7 +116,7 @@ registry_string_set(HKEY hkey, char *sub, char *x)
 {
 	DWORD   err;
 
-	err = RegSetValueEx(hkey, sub, 0, REG_SZ, (BYTE *)x, strlen(x) + 1);
+	err = RegSetValueEx(hkey, sub, 0, REG_SZ, (BYTE *)x, (DWORD)strlen(x) + 1);
 	if (err != ERROR_SUCCESS)
 		return -1;
